requirements.txt
telethon==1.24.0
requests==2.25.1
pillow

